requirements.txt
fastapi
uvicorn
mysql-connector-python
pydantic[email]
python-dotenv
groq
bcrypt
PyJWT
email-validator>=2.0.0
